+++ Assets/Scripts/Incidents/IncidentActions/BranchingAction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f6e9239aca2a05f49bc01ac10a577455
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
