+++ Assets/Scripts/Incidents/IncidentActions/IncidentActionBranch.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3f66faeea9ef96e47b455ed30eb9c723
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
