+++ Assets/Scripts/Incidents/IncidentActions/IncidentActionBranchWeightModifier.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a0bd01d29e595c44c87e4af9310bb778
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
